fix(multi-tenancy): fix backup/restore after banning namespace (#7442)
Fixes bunch of issues in multi-tenancy:

- If the incremental backup was taken after deleting a namespace, the data from the earlier 
    backups were restored. We fix this similar to the way we fix drop operations in backup.
- Mutation on deleted namespace should return, instead of retrying infinitely. This PR fixes it.
- Don't allow deletion of galaxy namespace.
- Extract namespace does not crash the cluster, when JWT is not passed correctly. 
   It returns an error instead.
- Adds a bunch of tests for the same.

+++ graphql/admin/namespace.go
@@ -19,11 +19,14 @@ package admin
 import (
 	"context"
 	"encoding/json"
+	"errors"
+	"fmt"
 	"strconv"
 
 	"github.com/dgraph-io/dgraph/edgraph"
 	"github.com/dgraph-io/dgraph/graphql/resolve"
 	"github.com/dgraph-io/dgraph/graphql/schema"
+	"github.com/dgraph-io/dgraph/x"
 )
 
 type addNamespaceInput struct {
@@ -62,7 +65,15 @@ func resolveDeleteNamespace(ctx context.Context, m schema.Mutation) (*resolve.Re
 	if err != nil {
 		return resolve.EmptyResult(m, err), false
 	}
+	// No one can delete the galaxy(default) namespace.
+	if uint64(req.NamespaceId) == x.GalaxyNamespace {
+		return resolve.EmptyResult(m, errors.New("Cannot delete default namespace.")), false
+	}
 	if err = (&edgraph.Server{}).DeleteNamespace(ctx, uint64(req.NamespaceId)); err != nil {
+		return resolve.EmptyResult(m, err), false
+	}
+	dropOp := "DROP_NS;" + fmt.Sprintf("%#x", req.NamespaceId)
+	if err = edgraph.InsertDropRecord(ctx, dropOp); err != nil {
 		return resolve.EmptyResult(m, err), false
 	}
 	return resolve.DataResult(
